v0.825 - Updated browser view and resolved warnings

+++ RawInputClient/Clock.cpp
@@ -12,6 +12,6 @@ CClock::~CClock()
 
 double CClock::GetTime()
 {
-	m_dbDuration = (std::clock() - m_clStartTime) / (double)CLOCKS_PER_SEC;	// Calculate current time
+	m_dbDuration = ((double)std::clock() - m_clStartTime) / (double)CLOCKS_PER_SEC;	// Calculate current time
 	return m_dbDuration;
 }
